feat(be, visit-activity): Reload children in activity filler.
When a child has not connected to an activity he/she has no connection to the actual visit so they cannot be reloaded when the filler reopened. In this commit I made a tmp data table for visits where we can save data such this.
Also I added visit info to the box-status so they can be connected now.

NM-111