fix(be, fe): Fix issues.
New app initializer should be added because auth guard now get the user from store and without the initer the user resolved later then the guard called. -> fixed

ReturnUrl did not work, needed to fix. -> fixed

When the login credentials were wrong the page gone a blank state where the user could not change the inputs. -> fixed

Needed an icon on the hospital visit details page to copy the link of the page -> added

Needed a checkbox on the creation of hospital visit to reopen the creator right after a successful save. It is really helpful when you want to add multiple visits. -> added

Event list page had just the list of cards view. It needed a table view. -> added

NM-129