feat: Replace old child storing technique to a new table.
Previously the children that was in a hospital visit and participate in an activity was stored in the activity record. Now in this commit I made them a new table and other to join the visited-child to the activity. Also made sql script to migrate the already recorded information.

NM-125