fix(fe): Box status list fixed to show only the related reports.
The box status list / manager showed all the status reports of the department where the visit happens. Now it only shows that status reports what are related to the actual visit. (Obviously in the department details page all status report of the department will be showed as before)

Added few style for visit details page.

NM-124