feat(be): Add support to filter participants on hospital visits.
Add a solution to convert FilterOptions in backend and then used it in HospitalVisitDao.
It is needed because if we just used TypeORM-s build in findAll function and add a filter on particitant.id,  the result will contain all the correct visits but the participant array in them will contain just the persons that satisfy the condition (not all participant). I think this behavior was not appropriate not even though it is logical from the viewpoint of SQL.

NM-138