feat: Take out user/userSettings from jwt.
Before that the jwt contained the user and userSettings. Now I changed and JWT includes only the userId and in the incoming request the JwtStrategy will load the current user to the request (with caching). User and UserSettings are provided in the response of the /login along with the accessToken.

NM-157