feat(fe): Add warning when there are activities when they should not be
In the visit details page:
It is possible to change the status of the visit to one of failed after you started the filling. In this situation the recorded stuff is probably invalid. Now a warning message appears and says the user should contact the coordinator. This message is just a warning because in rare cases it is valid.
Also, I added an info message when the status is STARTED to inform the user that the filling is not finalized.

NM-164