fix(fe): Add preloading to people selector.
In the people selector we just loaded the first 100 people because of simplicity. Now the amount of members of the prod system is greater than 100. I just changed the select to load all the available people and cache it. It could be work via lazy load too, but I wanted fewer requests. The memory consumption of people objects won't be so bad even if wonderfully grows 10x than it is now. And they are cached so the selector's filter will be much faster than in the lazy load version.

NM-181